renamed DefaultData to Fixtures
